requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain
chromadb
GitPython
beautifulsoup4
unstructured